fix(codegen): Swift compilation — result structs, closure returns, type mapping
Major Swift backend fixes:
- Generate result structs for multi-buffer protocol methods
  (AuthEncryptAuthEncryptResult, CipherAuthFinishAuthEncryptionResult,
   KemKemEncapsulateResult)
- Fix nested closure chain return propagation for multi-buffer methods
- Remove @objc from methods that throw + return non-bridgeable types
- Add String type mapping for string/char IR types
- Fix unsigned/integer type mapping to use sized types (UInt32, Int32)
- Handle disown-access class args with shallow_copy + pointer-to-pointer
- Resolve class_name="self" in return types to concrete class name
- Use interface name for result struct naming in implementation methods

Swift SPM compilation reaches deep into the wrapper code. Remaining:
1 protocol method has Void parameter type (Asn1Writer.reset) — a raw
byte pointer that needs type mapping refinement.

+++ wrappers/swift/VirgilCrypto/VirgilCryptoFoundation/Asn1Reader.swift
@@ -51,7 +51,7 @@ import VSCFoundation
     @objc func status() throws
 
     /// Get tag of the current ASN.1 element.
-    @objc func getTag() -> Int
+    @objc func getTag() -> Int32
 
     /// Get length of the current ASN.1 element.
     @objc func getLen() -> Int
@@ -61,42 +61,42 @@ import VSCFoundation
 
     /// Read ASN.1 type: TAG.
     /// Return element length.
-    @objc func readTag(tag: Int) -> Int
+    @objc func readTag(tag: Int32) -> Int
 
     /// Read ASN.1 type: context-specific TAG.
     /// Return element length.
     /// Return 0 if current position do not points to the requested tag.
-    @objc func readContextTag(tag: Int) -> Int
+    @objc func readContextTag(tag: Int32) -> Int
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readInt() -> Int
+    @objc func readInt() -> Int32
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readInt8() -> Int
+    @objc func readInt8() -> Int8
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readInt16() -> Int
+    @objc func readInt16() -> Int16
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readInt32() -> Int
+    @objc func readInt32() -> Int32
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readInt64() -> Int
+    @objc func readInt64() -> Int64
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readUint() -> Int
+    @objc func readUint() -> UInt32
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readUint8() -> Int
+    @objc func readUint8() -> UInt8
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readUint16() -> Int
+    @objc func readUint16() -> UInt16
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readUint32() -> Int
+    @objc func readUint32() -> UInt32
 
     /// Read ASN.1 type: INTEGER.
-    @objc func readUint64() -> Int
+    @objc func readUint64() -> UInt64
 
     /// Read ASN.1 type: BOOLEAN.
     @objc func readBool() -> Bool
